changes in seed data

+++ resources/reviews.py
@@ -1,12 +1,21 @@
-from flask import request
+from flask import request, jsonify
 from flask_restful import Resource
 from model import db, Review
+from flask_jwt_extended import jwt_required
 
-class ReviewRescource(Resource):
-    def get(self):
-        reviews = Review.query.all()
-        return [review.to_dict() for review in reviews], 200
+
+class ReviewResource(Resource):
+    @jwt_required()
+    def get(self, review_id):
+        # Retrieve a single review by ID
+        review = Review.query.get(review_id)
+        if not review:
+            return {'message': 'Review not found'}, 404
+        
+        return jsonify(review.to_dict())
     
+
+    @jwt_required()
     def post(self):
         data = request.get_json()
 
@@ -16,14 +25,25 @@ class ReviewRescource(Resource):
             rating = data['rating'],
             comment = data.get('comment', '')
         )
+        
+        
         db.session.add(new_review)
         db.session.commit()
         return new_review.to_dict(), 201
     
-    def delete(self, id):
-        review = Review.query.get_or_404(id)
+    @jwt_required()
+    def delete(self):
+        # Delete a review by ID
+        data = request.get_json()
+        review_id = data.get('review_id')
+        
+        review = Review.query.get(review_id)
+        if not review:
+            return {'message': 'Review not found'}, 404
+        
         db.session.delete(review)
         db.session.commit()
-        return {"message": "Review deleted successfully"}, 200
-    
+        
+        return {'message': 'Review deleted successfully'}, 200
+
